requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-WTF
Werkzeug==3.1.3
SQLAlchemy==2.0.36
python-dotenv==1.1.0
requests
